Remove unnecessary cast
Source: https://github.com/msupply-foundation/open-msupply/commit/4c8dd9523de53bc7bda429d5ca5bc50864d4fa3a

+++ server/service/src/invoice_line/stock_in_line/insert/generate.rs
@@ -99,7 +99,7 @@ fn generate_line(
     }: ItemRow,
     InvoiceRow { tax_percentage, .. }: InvoiceRow,
 ) -> InvoiceLineRow {
-    let total_before_tax = total_before_tax.unwrap_or(cost_price_per_pack * number_of_packs as f64);
+    let total_before_tax = total_before_tax.unwrap_or(cost_price_per_pack * number_of_packs);
     let total_after_tax = calculate_total_after_tax(total_before_tax, tax_percentage);
     InvoiceLineRow {
         id,
